#95 Migrate bound range structs to space header
