+++ motion_planning/libwheel/motion_planning/space.hpp
@@ -2,11 +2,45 @@
 #define LIBWHEEL_MOTION_PLANNING_SPACE_HPP
 
 #include <array>
+#include <cassert>
 #include <vector>
 
-#include "libwheel/motion_planning/bound_range.hpp"
-
 namespace wheel {
+
+/**
+ * @brief Wrapped type representing a BoundRange's lower bound
+ */
+struct LowerBound {
+    double value;
+};
+
+/**
+ * @brief Wrapped type representing a BoundRange's upper bound
+ */
+struct UpperBound {
+    double value;
+};
+
+/**
+ * @brief One-dimensional bounded range between [lower, upper] (inclusive)
+ *
+ * The lower value must be less than or equal to the upper value
+ */
+struct BoundRange {
+
+    /**
+     * @brief Construct a new BoundRange object
+     *
+     * @param l Lower range value
+     * @param u Upper range value
+     */
+    explicit BoundRange(LowerBound l, UpperBound u) : lower{l}, upper{u} {
+        assert((lower.value <= upper.value) && "lower bound must be <= upper bound");
+    }
+
+    LowerBound lower;
+    UpperBound upper;
+};
 
 /**
  * @brief Class representing a mathematical space (set with structure)
@@ -35,7 +69,6 @@ class Space {
      * @return const std::array<wheel::BoundRange, kDimensions>& Immutable reference to the space's bound ranges
      */
     auto get_bound_ranges() const -> const std::vector<wheel::BoundRange> & { return bound_ranges_; }
-
 
   private:
     std::vector<BoundRange> bound_ranges_;
